tasks/booking-service/src/requirements.txt
grpcio==1.75.1
protobuf==6.32.1
psycopg[binary]==3.2.10
requests==2.32.5
kafka-python==2.2.15